⚡️ :: ChallengeDTO 분리 & toDomain() 작성
